refactor: move data to follow project structure
