+++ src/error.rs
@@ -2,8 +2,10 @@ use axum::Json;
 use axum::http::StatusCode;
 use axum::http::header::WWW_AUTHENTICATE;
 use axum::response::{IntoResponse, Response};
+use serde::Serialize;
 use std::borrow::Cow;
 use std::collections::HashMap;
+use std::time::SystemTimeError;
 use thiserror::Error;
 
 /// A common error type that can be used throughout the API.
@@ -72,6 +74,64 @@ impl IntoResponse for Error {
             )
                 .into_response(),
             _ => (self.status_code(), self.to_string()).into_response(),
+        }
+    }
+}
+
+#[derive(Debug, Serialize)]
+pub struct ErrorResponse {
+    pub error: String,
+    pub message: String,
+}
+
+#[derive(thiserror::Error, Debug)]
+pub enum ServiceError {
+    #[error("Database error: {0}")]
+    DatabaseError(#[from] sqlx::Error),
+    #[error("System Time Error: {0}")]
+    SystemTimeError(#[from] SystemTimeError),
+    #[error("Invalid Project ID")]
+    InvalidProjectId(#[from] uuid::Error),
+}
+
+impl From<ServiceError> for (StatusCode, Json<ErrorResponse>) {
+    fn from(err: ServiceError) -> Self {
+        let (status, error_type, message) = match err {
+            ServiceError::DatabaseError(_) => (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                "database_error",
+                "Internal server error occurred",
+            ),
+            ServiceError::SystemTimeError(_) => (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                "system_time_error",
+                "Time went backwards",
+            ),
+            ServiceError::InvalidProjectId(_) => (
+                StatusCode::BAD_REQUEST,
+                "invalid_project_id",
+                "The provided project ID is invalid",
+            ),
+        };
+
+        (
+            status,
+            Json(ErrorResponse {
+                error: error_type.to_string(),
+                message: message.to_string(),
+            }),
+        )
+    }
+}
+
+impl From<ServiceError> for Error {
+    fn from(err: ServiceError) -> Self {
+        match err {
+            ServiceError::InvalidProjectId(_) => {
+                Error::unprocessable_entity([("project_id", "is invalid")])
+            }
+            ServiceError::DatabaseError(_) => Error::Forbidden,
+            ServiceError::SystemTimeError(_) => Error::Forbidden,
         }
     }
 }
