+++ src/lib.rs
@@ -3,7 +3,6 @@ pub mod db;
 pub mod error;
 pub mod http;
 pub mod middleware;
-pub mod services;
 pub mod telemetry;
 
 pub use config::*;
