--- src/services/mod.rs
@@ -1,3 +1,0 @@
-pub mod escrow;
-pub mod transaction;
-pub mod utils;
